while saving a document permissions for groups will be automatically assigned
